# Conflicts:

#	Passenger/app/src/main/java/com/fit/uet/passengerapp/Activity/fragments/FilterFragment.java

#	Passenger/app/src/main/java/com/fit/uet/passengerapp/adapter/ScheduleAdapter.java
#	Passenger/app/src/main/java/com/fit/uet/passengerapp/models/Coach.java

#	Passenger/app/src/main/java/com/fit/uet/passengerapp/models/CoachSchedule.java

+++ Passenger/app/src/main/java/com/fit/uet/passengerapp/models/Coach.java
@@ -15,7 +15,7 @@ public class Coach {
 
     public boolean hasWater;
     public boolean hasWifi;
-    public int kind;
+    public long kind;
     public String timeStart;
